requirements.txt
flask
flask-restx
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
httplib2==0.22.0
oauth2client==4.1.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.0
click==8.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2